requirements.txt
# FastAPI Backend Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.3.0
PyPDF2==3.0.1
python-docx==0.8.11
Pillow==10.1.0
pytesseract==0.3.10
watchdog==3.0.0

# Streamlit Frontend Requirements
streamlit==1.28.1
requests==2.31.0

# Additional dependencies
pydantic==2.5.0